Misc code cleanups to remove some common Eclipse warnings
git-svn-id: https://reprap.svn.sourceforge.net/svnroot/reprap/trunk/reprap/host@249 cb376a5e-1013-0410-a455-b6b1f9ac8223

+++ src/org/reprap/comms/snap/SNAPCommunicator.java
@@ -81,7 +81,7 @@ public class SNAPCommunicator implements Communicator {
 				System.out.print(device.getAddress().toString());
 				System.out.print(": ");
 				for(int i = 0; i < binaryMessage.length; i++)
-					System.out.print(Integer.toHexString((int)binaryMessage[i]) + " ");
+					System.out.print(Integer.toHexString(binaryMessage[i]) + " ");
 				System.out.println("");
 			}
 			sendRawMessage(packet);
